Add beginning of twitch.go
This is a go script that gets a list of streamers currently playing
Fortnite and will send its results to job-controller. It currently
seems pretty slow and could probably perform much faster by making
one batch call to GetUsersByID rather than a call for each user.

Writing this on a plane so fucks given on that are low and it is a
problem for tomorrow.

Signed-off-by: Cullen Taylor <mctaylor@us.ibm.com>