Add twitch.go to handle getting list of streams
This breaks job-controller for now since the client_id environment
variable is not present in the docker image but will be present as
a kubernetes secret.

Adds a place for twitch functions (this should be the only one
however) to live.

Next step is to figure out go's redis client and start storing the
obtained streams list there.

Signed-off-by: Cullen Taylor <mctaylor@us.ibm.com>