Delete twitch.go
job-controller should just check for work stored in redis. If there
is work to do, then it will fire off a job which will spawn a set
of pods and check to see if that work is completed. If there is no
work to do, then do nothing until the next interval is hit.

The above behavior still needs to be implemented. This just removes
the cruft.

Signed-off-by: Cullen Taylor <mctaylor@us.ibm.com>