Remove types.go
There is no need for the Stream struct type anymore, as tracking
twitch streams does not need to be part of job-controller. The only
other type is the Client struct type which defines a Kubernetes
clientset, and that can just live in executor.go, as that is the
sole place it is needed.

Signed-off-by: Cullen Taylor <mctaylor@us.ibm.com>