Set propagation policy when deleting jobs
Kubernetes (1.9 and above), dependent objects are deleted by
default, and this is true when using kubectl. When deleting jobs
using the client-go api, you have to set the deletion propagation
policy when deleting resources in order to have their dependent
objects deleted as well.

I chose the Background deletion policy, as it's okay that the job
is deleted first then the pods are cleaned up afterward. The
Foreground deletion policy would block job deletion until all of
the dependent pods are cleaned up.

Signed-off-by: Cullen Taylor <mctaylor@us.ibm.com>